Resized Images
Resized images to significantly reduce the amount of data transferred.
Also, set the modal popup to display a larger image. However, this has
caused an increase in the number of resource requests due to the way
the HTML is structured. Need to understand how Foundation Reveal works
to fix this.